requirements.txt
torch
torchvision
torchaudio
ultralytics
numpy
pandas
pyyaml
matplotlib
jupyterlab
ipykernel
opencv-python
tqdm
scikit-learn
